third_party/WebKit/LayoutTests/media/controls/controls-overlay-cast-button-autoplay-muted.html

<!DOCTYPE html>
<title>media controls overlay cast button for autoplay muted element</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../media-file.js"></script>
<script src="../media-controls.js"></script>
<video autoplay muted></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    video.src = findMediaFile("video", "../content/test");

    video.onloadedmetadata = t.step_func_done(function() {
        // Should not have a cast button by default
        var button = overlayCastButton(video);
        assert_false(isControlVisible(button), "button should not be visible with no cast devices");

        // Pretend we have a cast device
        internals.mediaPlayerRemoteRouteAvailabilityChanged(video, true);

        // Now should have cast button
        assert_false(isControlVisible(button), "button should not be visible for autoplay muted");

        // Remove cast device - cast button should go away
        internals.mediaPlayerRemoteRouteAvailabilityChanged(video, false);
    });
}, 'autoplay muted video');
</script>
